recommendation/requirements.txt
flask
flask-cors
psycopg2-binary
schedule
tensorflow==2.13